+++ src/digi_edit/views/__init__.py
@@ -8,4 +8,5 @@ def root(request):
 
 @view_config(route_name='ui', renderer='digi_edit:templates/ui.jinja2')
 def ui(request):
+    request.response.headers['Cache-Control'] = 'no-store max-age=0'
     return {}